+++ infra/environments/dev/main.tf
@@ -10,6 +10,7 @@ module "relayworks" {
   control_plane_image       = var.control_plane_image
   sync_worker_image         = var.sync_worker_image
   control_plane_api_client_id = var.control_plane_api_client_id
+  console_client_id           = var.console_client_id
 
   tags = {
     application = "relayworks"