+++ infra/modules/relayworks-platform/variables.tf
@@ -30,6 +30,7 @@ variable "sync_worker_image" {
   type = string
 }
 variable "control_plane_api_client_id" { type = string }
+variable "console_client_id" { type = string }
 
 variable "tags" {
   type = map(string)